+++ ql/InterestRateModelling/grid.hpp
@@ -49,7 +49,7 @@ namespace QuantLib {
 
         class Grid : public Array {
           public:
-            Grid(unsigned int gridPoints,
+            Grid(size_t gridPoints,
                  double initialCenter,
                  double strikeCenter,
                  Time residualTime,
@@ -59,7 +59,7 @@ namespace QuantLib {
                 initialize(gridPoints, initialCenter, strikeCenter,
                     residualTime, timeDelay, model);
             }
-            Grid(unsigned int gridPoints,
+            Grid(size_t gridPoints,
                  double initialCenter,
                  double strikeCenter,
                  Time residualTime,
@@ -73,23 +73,23 @@ namespace QuantLib {
             double xMin() {return (*this)[0];}
             double xMax() {return (*this)[size()-1];}
             double dx() { return dx_;}
-            unsigned int index() const {return index_;}
-            unsigned int safeGridPoints(unsigned int gridPoints,
+            size_t index() const {return index_;}
+            size_t safeGridPoints(unsigned int gridPoints,
               Time residualTime) const;
           private:
-            void initialize(unsigned int gridPoints,
+            void initialize(size_t gridPoints,
                  double initialCenter,
                  double strikeCenter,
                  Time residualTime,
                  Time timeDelay,
                  const Handle<OneFactorModel>& model);
             double dx_;
-            unsigned int index_;
+            size_t index_;
 
         };
 
 
-        inline void Grid::initialize(unsigned int gridPoints,
+        inline void Grid::initialize(size_t gridPoints,
           double initialCenter, double strikeCenter,
           Time residualTime, Time timeDelay,
           const Handle<OneFactorModel>& model) {
@@ -113,13 +113,13 @@ namespace QuantLib {
             dx_ = (xMax - xMin)/(size()-1);
             for (unsigned j=0; j<size(); j++)
                 (*this)[j] = xMin + j*dx_;
-            index_ = (unsigned int)((initialCenter - xMin)/dx_ + 0.5);
+            index_ = size_t(((initialCenter - xMin)/dx_ + 0.5));
         }
 
-        inline unsigned int Grid::safeGridPoints(
-            unsigned int gridPoints, Time residualTime) const {
+        inline size_t Grid::safeGridPoints(
+            size_t gridPoints, Time residualTime) const {
             return QL_MAX(gridPoints, residualTime>1.0 ?
-                static_cast<unsigned int>(
+                static_cast<size_t>(
                     (QL_NUM_OPT_MIN_GRID_POINTS +
                     (residualTime-1.0) *
                     QL_NUM_OPT_GRID_POINTS_PER_YEAR))
